Expand filenames in `:scope` like agenda files (#39)
* Expand filenames in scope like agenda files

The :scope tag cannot contain files with multiple subdirectories because
org-kanban.el and org itself keeps state (the working directory) when
checking whether (agenda) files exists.  Since org-mode has an interface
to add agenda files resulting in the files to be added in the form of
absolute paths, this commit suggests to treat the parameters of :scope
in the same way.  Added benefit is that it allows you to use a directory
in the :scope tag which results in all org files to be added to the
kanban table.

Co-authored-by: Christian Koestlin <christian.koestlin@gmail.com>